@@ -3,17 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.3.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^4.2.2",
+    "bootstrap": "^5.2.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "^5.0.1",
-    "reactstrap": "^6.3.0",
-    "rimraf": "^2.6.2"
+    "reactstrap": "^9.1.9"
   },
   "devDependencies": {
-    "ajv": "^8.12.0",
-    "cross-env": "^5.2.0"
+    "cross-env": "7.0.3",
+    "rimraf": "5.0.0"
   },
   "scripts": {
     "start": "rimraf ./build && react-scripts start",
